fix search loop
